requirements.txt
# Web-фреймворк и REST API
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-environ>=0.11.2
# JWT-аутентификация
djangorestframework-simplejwt>=5.3,<6.0

# База данных PostgreSQL (для продакшена)
psycopg2-binary>=2.9,<3.0
django_celery_beat
# Работа с .env файлами
python-dotenv>=1.0,<2.0

# Сервер для продакшена
gunicorn>=22.0,<23.0

# Мониторинг Prometheus
django-prometheus>=2.3,<3.0  # поддержка Django >=4.2 и Python³⁹+ :contentReference[oaicite:1]{index=1}

# Документация API
drf-spectacular>=0.27,<1.0

# Фильтрация, поиск задач
django-filter>=24.2,<25.0

# Тестирование
pytest>=8.2,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
coverage>=7.5,<8.0

# Планируемое: Celery (для отложенных задач)
celery>=5.4,<6.0
redis>=5.0,<6.0

whitenoise